+++ applications/ionMobility/RS-DMSSimplifiedSim/RS-DMSSimplifiedSim.cpp
@@ -178,7 +178,7 @@ int main(int argc, const char * argv[]) {
                 particle->setFloatAttribute(key_ChemicalIndex, substIndex);
 
                 particlesPtrs.push_back(particle.get());
-                rsSim.addParticle(particle.get(), static_cast<int>(nParticlesTotal));
+                rsSim.addParticle(particle.get(), nParticlesTotal);
                 particles.push_back(std::move(particle));
                 trajectoryAdditionalParams.emplace_back(std::vector<double>(1));
                 nParticlesTotal++;
@@ -298,7 +298,7 @@ int main(int argc, const char * argv[]) {
             reactionConditions.electricField = fieldMagnitude;
 
             for (unsigned int i = 0; i < nParticlesTotal; i++) {
-                bool reacted = rsSim.react(static_cast<int>(i), reactionConditions, dt_s);
+                bool reacted = rsSim.react(i, reactionConditions, dt_s);
 
                 if (reacted){
                     //we had an reaction event: Update the chemical species for the trajectory
